+++ CHANGES.rst
@@ -407,6 +407,8 @@ Bug Fixes
 
 - ``astropy.utils``
 
+  - Fixed latex representation of function units. [#4563]
+
 - ``astropy.visualization``
 
   - Fixed ``fits2bitmap`` script to allow ext flag to contain extension
@@ -1174,6 +1176,8 @@ Bug Fixes
 - ``astropy.time``
 
 - ``astropy.units``
+
+  - Fixed latex representation of function units. [#4563]
 
 - ``astropy.utils``